+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 - Refactor `default_backend` to return a `Box<dyn Backend>`.
 - Check the `DIALOG` and `DISPLAY` environment variables in `default_backend`.
+- Add the `Stdio` backend.
 
 # v0.1.1 (2019-01-11)
 - Add the `Password` dialog box.